Rename Image Archive to Media Archive
As per refined requirements. Changing the most import things and UI
elements.
