+++ src/Loader.php
@@ -86,8 +86,8 @@ final class Loader
 
             // Load P4 Control Panel only on Dashboard page.
             $this->default_services[] = ControlPanel::class;
-            $this->default_services[] = ImageArchive\UiIntegration::class;
-            $this->default_services[] = ImageArchive\Rest::class;
+            $this->default_services[] = MediaArchive\UiIntegration::class;
+            $this->default_services[] = MediaArchive\Rest::class;
 
             // Load P4 Metaboxes only when adding/editing a new Page/Post/Campaign.
             if ('post-new.php' === $pagenow || 'post.php' === $pagenow) {
