+++ webpack.config.js
@@ -30,7 +30,7 @@ module.exports = {
     "navigation-bar-light": './assets/src/scss/partials/navigation-bar-light.scss',
     "gravity-forms": './assets/src/scss/layout/_gravity-forms.scss',
     new_identity_styles: './assets/src/scss/new-identity/style.scss',
-    archive_picker: './assets/src/js/archive_picker.js',
+    media_archive: './assets/src/js/media_archive.js',
     "lite-yt-embed": './node_modules/lite-youtube-embed/src/lite-yt-embed.js',
     menu_editor: './assets/src/js/menu_editor.js',
   },
